FIxed warning

+++ Controls4U/SliderCtrlX.cpp
@@ -552,7 +552,7 @@ int SliderCtrlX::SliderToClient(int v) const
 
 SliderCtrlX& SliderCtrlX::SetMajorTicks(int n) { 
 	if (IsNull(n))
-		m_nMajorTicks = GetRangeMajorUnits(m_nMin, m_nMax);
+		m_nMajorTicks = (int)GetRangeMajorUnits(m_nMin, m_nMax);
 	else
 		m_nMajorTicks = n; 
 	return *this;
